requirements.txt
asgiref==3.7.2
boto3==1.34.21
botocore==1.34.21
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==41.0.7
decorator==4.4.2
dj-database-url==2.1.0
Django==5.0
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-storages==1.14.2
djangorestframework==3.14.0
gunicorn==21.2.0
idna==3.6
imageio==2.33.1
imageio-ffmpeg==0.4.9
jmespath==1.0.1
numpy==1.26.3
packaging==23.2
Pillow==10.1.0
proglog==0.1.10
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
s3transfer==0.10.0
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.0.7
whitenoise==6.6.0
